+++ core/communication/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/communication/impl/MetisServiceImpl.kt
@@ -120,7 +120,7 @@ class MetisServiceImpl(
      */
     override suspend fun getPost(
         metisContext: MetisContext,
-        serverSidePostId: Int,
+        serverSidePostId: Long,
         serverUrl: String,
         authToken: String
     ): NetworkResponse<StandalonePost> {
