+++ core/datastore/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/datastore/model/metis/MetisContext.kt
@@ -30,7 +30,7 @@ sealed class MetisContext(
 
     @Serializable
     @SerialName("lecture")
-    data class Lecture(override val courseId: Long, val lectureId: Int) :
+    data class Lecture(override val courseId: Long, val lectureId: Long) :
         MetisContext(
             listOf("courses", courseId.toString(), "lectures", lectureId.toString()),
             "posts",
