+++ core/model/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/model/metis/BasePost.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 sealed class BasePost {
-    abstract val id: Int?
+    abstract val id: Long?
     abstract val author: User?
     abstract val authorRole: UserRole?
     abstract val creationDate: Instant?
